Move `getValidator` to database helpers :wrench:
