+++ server/common/database/helpers.js
@@ -1,7 +1,11 @@
 'use strict';
 
+const inRange = require('lodash/inRange');
+const notEmpty = input => input.length > 0;
+
 module.exports = {
-  alterEnum
+  alterEnum,
+  getValidator
 };
 
 // NOTE: Enables safe altering of ENUM values:
@@ -12,4 +16,16 @@ function alterEnum(queryInterface, Sequelize, table, column, options = {}) {
   return queryInterface.changeColumn(table, column, str)
     .then(() => sequelize.query(QueryGenerator.pgEnumDrop(table, column)))
     .then(() => queryInterface.changeColumn(table, column, options));
+}
+
+function getValidator(Model, attribute) {
+  return function validate(input) {
+    const validator = Model.prototype.validators[attribute];
+    if (!validator || !validator.len) {
+      return notEmpty(input) || `"${attribute}" can not be empty`;
+    }
+    const [min, max] = validator.len;
+    return inRange(input.length, min, max) ||
+      `"${attribute}" must be between ${min} and ${max} characters long`;
+  };
 }
